On branch res/mtree_compare changed: 	modified:   cluster/src/main/java/org/apache/iotdb/cluster/query/ClusterPlanExecutor.java 	modified:   example/session/pom.xml 	new file:   example/session/src/main/java/org/apache/iotdb/MTreeBenchSession.java 	modified:   example/session/src/main/java/org/apache/iotdb/pathgenerator/L1Prefix.java 	modified:   example/session/src/main/java/org/apache/iotdb/pathgenerator/L2Prefix.java 	modified:   example/session/src/main/java/org/apache/iotdb/pathgenerator/MeasurementName.java 	modified:   example/session/src/main/java/org/apache/iotdb/pathgenerator/PathGenerator.java 	modified:   pom.xml 	modified:   server/pom.xml 	modified:   server/src/main/java/org/apache/iotdb/db/conf/IoTDBConfig.java 	modified:   server/src/main/java/org/apache/iotdb/db/metadata/MManager.java 	modified:   server/src/main/java/org/apache/iotdb/db/qp/executor/PlanExecutor.java 	modified:   server/src/main/java/org/apache/iotdb/db/service/IoTDB.java 	modified:   service-rpc/src/main/java/org/apache/iotdb/rpc/AutoResizingBuffer.java 	modified:   session/src/main/java/org/apache/iotdb/session/Session.java
 Untracked files:
	mlog_sample/

+++ cluster/src/main/java/org/apache/iotdb/cluster/query/ClusterPlanExecutor.java
@@ -127,7 +127,7 @@ public class ClusterPlanExecutor extends PlanExecutor {
   }
 
   @Override
-  protected int getDevicesNum(PartialPath path, boolean isPrefixMatch) throws MetadataException {
+  protected long getDevicesNum(PartialPath path, boolean isPrefixMatch) throws MetadataException {
     // adapt to prefix match of IoTDB v0.12
     return getDevicesNum(path)
         + (isPrefixMatch
